+++ src/www/ui_components/server_card/index.ts
@@ -19,16 +19,6 @@ import {ServerConnectionState} from "./server_connection_viz";
 
 export * from "./server_connection_viz";
 
-export interface ServerCardModel {
-  disabled: boolean;
-  errorMessageId: string;
-  isOutlineServer: boolean;
-  address: string;
-  id: string;
-  name: string;
-  state: ServerConnectionState;
-}
-
 @customElement("server-card")
 export class ServerCard extends LegacyElementMixin(PolymerElement) {
   static template = html`
