+++ src/www/ui_components/server_card/server_connection_viz/index.ts
@@ -14,15 +14,9 @@
 import {computed, customElement, property} from "@polymer/decorators";
 import {html, PolymerElement} from "@polymer/polymer";
 import {LegacyElementMixin} from "@polymer/polymer/lib/legacy/legacy-element-mixin";
+import {ServerConnectionState} from "./types";
 
-export enum ServerConnectionState {
-  INITIAL = "INITIAL",
-  CONNECTING = "CONNECTING",
-  CONNECTED = "CONNECTED",
-  RECONNECTING = "RECONNECTING",
-  DISCONNECTING = "DISCONNECTING",
-  DISCONNECTED = "DISCONNECTED",
-}
+export * from "./types";
 
 @customElement("server-connection-viz")
 export class ServerConnectionViz extends LegacyElementMixin(PolymerElement) {
