+++ src/www/ui_components/server_list/types.ts
@@ -1,0 +1,11 @@
+import {ServerConnectionState} from "../server_card";
+
+export interface ServerListItem {
+  disabled: boolean;
+  errorMessageId: string;
+  isOutlineServer: boolean;
+  address: string;
+  id: string;
+  name: string;
+  state: ServerConnectionState;
+}
